[VEN-679] Add ts linting

+++ spec/certora/contracts/CDaiDelegateCertora.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.10;
 import "../../../contracts/CDaiDelegate.sol";
 
 contract CDaiDelegateCertora is CDaiDelegate {
-    function getCashOf(address account) public view returns (uint) {
+    function getCashOf(address account) public view returns (uint256) {
         return EIP20Interface(underlying).balanceOf(account);
     }
 }
