+++ spec/certora/contracts/CErc20DelegateCertora.sol
@@ -8,26 +8,38 @@ import "./VTokenCollateral.sol";
 contract VBep20DelegateCertora is VBep20Delegate {
     VTokenCollateral public otherToken;
 
-    function mintFreshPub(address minter, uint mintAmount) public returns (uint) {
-        (uint error,) = mintFresh(minter, mintAmount);
+    function mintFreshPub(address minter, uint256 mintAmount) public returns (uint256) {
+        (uint256 error, ) = mintFresh(minter, mintAmount);
         return error;
     }
 
-    function redeemFreshPub(address payable redeemer, uint redeemTokens, uint redeemUnderlying) public returns (uint) {
+    function redeemFreshPub(
+        address payable redeemer,
+        uint256 redeemTokens,
+        uint256 redeemUnderlying
+    ) public returns (uint256) {
         return redeemFresh(redeemer, redeemTokens, redeemUnderlying);
     }
 
-    function borrowFreshPub(address payable borrower, uint borrowAmount) public returns (uint) {
+    function borrowFreshPub(address payable borrower, uint256 borrowAmount) public returns (uint256) {
         return borrowFresh(borrower, borrowAmount);
     }
 
-    function repayBorrowFreshPub(address payer, address borrower, uint repayAmount) public returns (uint) {
-        (uint error,) = repayBorrowFresh(payer, borrower, repayAmount);
+    function repayBorrowFreshPub(
+        address payer,
+        address borrower,
+        uint256 repayAmount
+    ) public returns (uint256) {
+        (uint256 error, ) = repayBorrowFresh(payer, borrower, repayAmount);
         return error;
     }
 
-    function liquidateBorrowFreshPub(address liquidator, address borrower, uint repayAmount) public returns (uint) {
-        (uint error,) = liquidateBorrowFresh(liquidator, borrower, repayAmount, otherToken);
+    function liquidateBorrowFreshPub(
+        address liquidator,
+        address borrower,
+        uint256 repayAmount
+    ) public returns (uint256) {
+        (uint256 error, ) = liquidateBorrowFresh(liquidator, borrower, repayAmount, otherToken);
         return error;
     }
 }
